requirements.txt
Django==5.0
Pillow==10.0.0
django-ckeditor-5==0.2.17
gunicorn==21.2.0
sqlparse==0.5.3
packaging==24.2
asgiref==3.8.1
whitenoise==6.8.2
django-storages==1.14.4 
google-cloud-storage==3.0.0 
django-extensions==3.2.3